fix: Competition typing in Competition pages
This commit add the typing of Competition to the Competition pages in the frontend_profile
Because of the repeated name for the Interface and the Component, the Component was renamed to CompetitionItem.
