+++ frontend_profile/src/pages/MyCompetitionPage/MyCompetitionPage.tsx
@@ -47,7 +47,7 @@ export default function MyCompetitionPage() {
           sx={{ display: "flex", justifyContent: "center" }}
         />
         {competitions?.length === 0 && <h2>沒有更多比賽了喲 ;(</h2>}
-        {isLoadingCompetitions || isLoadingUid || !uid ? (
+        {isLoadingCompetitions || isLoadingUid || !uid || !competitions ? (
           <p>loading...</p>
         ) : (
           <CompetitionList competitions={competitions} uid={uid} />
