+++ frontend_profile/src/pages/MyCompetitionPage/components/CompetitionList.tsx
@@ -1,6 +1,5 @@
+import { CompetitionItem } from "./CompetitionItem";
 import { Competition } from "../../../util/QueryHooks/types/Competition";
-import { Competition } from "./Competition";
-
 interface Props {
   competitions: Competition[];
   uid: number;
@@ -8,6 +7,6 @@ interface Props {
 
 export const CompetitionList = ({ competitions, uid }: Props) => {
   return competitions?.map((competition: any) => (
-    <Competition competition={competition} uid={uid} key={competition.id} />
+    <CompetitionItem competition={competition} uid={uid} key={competition.id} />
   ));
 };
