+++ frontend_profile/src/pages/RecentCompetitionPage/RecentCompetitionPage.tsx
@@ -48,7 +48,7 @@ export default function RecentCompetitionPage() {
           sx={{ display: "flex", justifyContent: "center" }}
         />
         {competitions?.length === 0 && <h2>沒有更多比賽了喲 ;(</h2>}
-        {isLoadingCompetitions || isLoadingUid || !uid ? (
+        {isLoadingCompetitions || isLoadingUid || !uid || !competitions ? (
           <p>loading...</p>
         ) : (
           <CompetitionList competitions={competitions} uid={uid} />
