+++ frontend_profile/src/pages/RecentCompetitionPage/components/CompetitionList.tsx
@@ -1,12 +1,12 @@
-import { Competition } from "./Competition";
-
+import { CompetitionItem } from "./CompetitionItem";
+import { Competition } from "../../../util/QueryHooks/types/Competition";
 interface Props {
-  competitions: any;
+  competitions: Competition[];
   uid: number;
 }
 
 export const CompetitionList = ({ competitions, uid }: Props) => {
   return competitions?.map((competition: any) => (
-    <Competition competition={competition} uid={uid} key={competition.id} />
+    <CompetitionItem competition={competition} uid={uid} key={competition.id} />
   ));
 };
